Split guard extraction from constraints
